+++ src/desktop/src/ui/global/ErrorLog.js
@@ -70,7 +70,7 @@ export class ErrorLogComponent extends React.PureComponent {
                                                     {formatTimeAs.dayMonthYearHoursMinutes(
                                                         navigator.language,
                                                         detectedTimezone,
-                                                        entry.date,
+                                                        entry.time,
                                                     )}
                                                 </strong>{' '}
                                                 {entry.error}
@@ -102,7 +102,4 @@ const mapDispatchToProps = {
     clearLog,
 };
 
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps,
-)(withTranslation()(ErrorLogComponent));
+export default connect(mapStateToProps, mapDispatchToProps)(withTranslation()(ErrorLogComponent));