+++ src/desktop/src/ui/global/Polling.js
@@ -23,7 +23,6 @@ import {
     setPollFor,
     promoteTransfer,
     getAccountInfoForAllAccounts,
-    fetchMeta as fetchMoonPayMeta,
 } from 'actions/polling';
 import { retryFailedTransaction } from 'actions/transfers';
 
@@ -58,8 +57,6 @@ class Polling extends React.PureComponent {
         fetchNodeList: PropTypes.func.isRequired,
         /** @ignore */
         promoteTransfer: PropTypes.func.isRequired,
-        /** @ignore */
-        fetchMoonPayMeta: PropTypes.func.isRequired,
         /** Bundle hashes for failed transactions categorised by account name & type */
         failedBundleHashes: PropTypes.shape({
             name: PropTypes.string,
@@ -76,8 +73,6 @@ class Polling extends React.PureComponent {
         /** @ignore */
         isPollingAccountInfo: PropTypes.bool.isRequired,
         /** @ignore */
-        isPollingMoonPayMeta: PropTypes.bool.isRequired,
-        /** @ignore */
         isAutoPromoting: PropTypes.bool.isRequired,
         /** @ignore */
         isTransitioning: PropTypes.bool.isRequired,
@@ -93,8 +88,6 @@ class Polling extends React.PureComponent {
         isSendingTransfer: PropTypes.bool.isRequired,
         /** @ignore */
         isFetchingAccountInfo: PropTypes.bool.isRequired,
-        /** @ignore */
-        isAuthenticatedForMoonPay: PropTypes.bool.isRequired,
     };
 
     state = {
@@ -136,7 +129,6 @@ class Polling extends React.PureComponent {
             nodeList: this.props.fetchNodeList,
             accountInfo: this.fetchLatestAccountInfo,
             broadcast: this.retryFailedTransaction,
-            moonpayMeta: this.fetchMoonPayMeta,
         };
 
         dict[service] ? dict[service]() : this.props.setPollFor(this.props.allPollingServices[0]);
@@ -148,23 +140,6 @@ class Polling extends React.PureComponent {
             [selectedAccountName, ...filter(accountNames, (name) => name !== selectedAccountName)],
             Electron.notify,
         );
-    };
-
-    /**
-     * Fetch MoonPay transaction history (if user is authenticated)
-     *
-     * @method fetchMoonPayTransactions
-     *
-     * @returns {void}
-     */
-    fetchMoonPayMeta = () => {
-        const { isAuthenticatedForMoonPay } = this.props;
-
-        if (isAuthenticatedForMoonPay) {
-            this.props.fetchMoonPayMeta();
-        } else {
-            this.moveToNextPollService();
-        }
     };
 
     /**
@@ -230,8 +205,7 @@ class Polling extends React.PureComponent {
             this.props.isPollingChartData ||
             this.props.isPollingMarketData ||
             this.props.isPollingAccountInfo ||
-            this.props.isAutoPromoting ||
-            this.props.isPollingMoonPayMeta;
+            this.props.isAutoPromoting;
 
         return isAlreadyDoingSomeHeavyLifting || isAlreadyPollingSomething;
     }
@@ -249,7 +223,6 @@ const mapStateToProps = (state) => ({
     isPollingMarketData: state.polling.isFetchingMarketData,
     isPollingAccountInfo: state.polling.isFetchingAccountInfo,
     isAutoPromoting: state.polling.isAutoPromoting,
-    isPollingMoonPayMeta: state.polling.isFetchingMoonPayMeta,
     isSyncing: state.ui.isSyncing,
     addingAdditionalAccount: isSettingUpNewAccount(state),
     isGeneratingReceiveAddress: state.ui.isGeneratingReceiveAddress,
@@ -264,7 +237,6 @@ const mapStateToProps = (state) => ({
     isRetryingFailedTransaction: state.ui.isRetryingFailedTransaction,
     failedBundleHashes: getFailedBundleHashes(state),
     password: state.wallet.password,
-    isAuthenticatedForMoonPay: state.exchanges.moonpay.isAuthenticated,
 });
 
 const mapDispatchToProps = {
@@ -274,7 +246,6 @@ const mapDispatchToProps = {
     promoteTransfer,
     getAccountInfoForAllAccounts,
     retryFailedTransaction,
-    fetchMoonPayMeta,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Polling);