+++ src/mobile/src/ui/components/ModalComponent.js
@@ -20,7 +20,6 @@ import Checksum from 'ui/components/ChecksumModal';
 import QrScanner from 'ui/components/QrScanner';
 import Print from 'ui/components/PrintModal';
 import BiometricInfo from 'ui/components/BiometricInfoModal';
-import MoonPayPurchaseDetails from 'ui/components/exchanges/MoonPay/PurchaseDetailsModal';
 import NotificationLog from 'ui/components/NotificationLogModal';
 import { isAndroid, isIPhoneX } from 'libs/device';
 import { getThemeFromState } from 'shared-modules/selectors/global';
@@ -57,7 +56,6 @@ const MODAL_CONTENT = {
     biometricInfo: BiometricInfo,
     notificationLog: NotificationLog,
     checksum: Checksum,
-    moonpayPurchaseDetails: MoonPayPurchaseDetails,
 };
 
 const fadeInUpCustom = {
@@ -196,8 +194,5 @@ export default function withSafeAreaView(WrappedComponent) {
         toggleModalActivity,
     };
 
-    return connect(
-        mapStateToProps,
-        mapDispatchToProps,
-    )(EnhancedComponent);
+    return connect(mapStateToProps, mapDispatchToProps)(EnhancedComponent);
 }