+++ src/mobile/src/ui/views/wallet/Receive.js
@@ -793,7 +793,6 @@ const mapStateToProps = (state) => ({
     usdPrice: state.marketData.usdPrice,
     conversionRate: state.settings.conversionRate,
     hadErrorGeneratingNewAddress: state.ui.hadErrorGeneratingNewAddress,
-    isViewingMoonpayPurchases: state.ui.isViewingMoonpayPurchases,
 });
 
 const mapDispatchToProps = {