+++ src/shared/__tests__/reducers/ui.spec.js
@@ -38,8 +38,6 @@ describe('Reducer: ui', () => {
                 hadErrorGeneratingNewAddress: false,
                 isKeyboardActive: false,
                 animateChartOnMount: true,
-                isRecoveringFunds: false,
-                isViewingMoonpayPurchases: false,
                 CDAContent: {},
             };