+++ src/shared/containers/components/Chart.js
@@ -23,8 +23,6 @@ export default function withChartData(ChartComponent) {
             setChartCurrency: PropTypes.func.isRequired,
             t: PropTypes.func.isRequired,
             theme: PropTypes.object.isRequired,
-            /** @ignore */
-            isAuthenticated: PropTypes.bool.isRequired,
             history: PropTypes.object.isRequired,
         };
 
@@ -91,7 +89,7 @@ export default function withChartData(ChartComponent) {
         };
 
         render() {
-            const { history, isAuthenticated, marketData, settings, theme, t } = this.props;
+            const { history, marketData, settings, theme, t } = this.props;
 
             const currencyData = get(marketData.chartData, settings.chartCurrency.toLowerCase());
             const rawData = get(currencyData, settings.chartTimeframe) || [];
@@ -103,7 +101,6 @@ export default function withChartData(ChartComponent) {
             });
             const chartProps = {
                 history,
-                isAuthenticatedForMoonPay: isAuthenticated,
                 setCurrency: this.changeCurrency,
                 setTimeframe: this.changeTimeframe,
                 getPriceFormat: this.getPriceFormat,
@@ -138,7 +135,6 @@ export default function withChartData(ChartComponent) {
         marketData: state.marketData,
         settings: state.settings,
         theme: getThemeFromState(state),
-        isAuthenticated: state.exchanges.moonpay.isAuthenticated,
     });
 
     const mapDispatchToProps = {