+++ src/shared/icons/icons.js
@@ -22,8 +22,6 @@ module.exports = {
     vault: '📦',
     advanced: '🔬',
     mode: '🔣',
-    moonpay: '🐧',
-    moonpayEmblem: '🌙',
     logout: '🖐',
     currency: '💵',
     send: '📤',